realtabbench/agent_eval/requirements.txt
tablegpt-agent
aiofiles
tqdm
pydantic >= 2.0
pydantic-settings >= 2.0
python-dotenv
pyyaml
ipython
ipykernel
